pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^2.0.3"
python-dotenv = "^0.14.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
